pull out lh5concat into a seperate function
