+++ tests/test_cli.py
@@ -23,7 +23,7 @@ def test_lh5concat(lgnd_test_data, tmptestdir):
         "lh5/prod-ref-l200/generated/tier/raw/phy/p03/r001/l200-p03-r001-phy-20230322T170202Z-tier_raw.lh5"
     )
     outfile = f"{tmptestdir}/out.lh5"
-    cli.lh5concat(["--output", outfile, "--", infile1, infile2])
+    cli.lh5concat_cli(["--output", outfile, "--", infile1, infile2])
 
     assert lh5.ls(outfile) == [
         "ch1057600",
@@ -100,13 +100,13 @@ def test_lh5concat(lgnd_test_data, tmptestdir):
         infile2,
     ]
 
-    cli.lh5concat(arg_list)
+    cli.lh5concat_cli(arg_list)
     assert lh5.ls(outfile) == [
         "ch1057600",
     ]
 
     arg_list[5] = "ch1057600/raw/waveform/values"
-    cli.lh5concat(arg_list)
+    cli.lh5concat_cli(arg_list)
     assert lh5.ls(outfile, "ch1057600/raw/waveform/") == [
         "ch1057600/raw/waveform/values",
     ]
@@ -117,7 +117,7 @@ def test_lh5concat(lgnd_test_data, tmptestdir):
     arg_list[4] = "--exclude"
     arg_list[5] = "ch1057600/raw/waveform/values"
 
-    cli.lh5concat(arg_list)
+    cli.lh5concat_cli(arg_list)
     assert lh5.ls(outfile) == [
         "ch1057600",
         "ch1059201",
@@ -162,7 +162,7 @@ def test_lh5concat(lgnd_test_data, tmptestdir):
     store.write(struct2, "stp", infile2, wo_mode="overwrite_file")
 
     outfile = f"{tmptestdir}/concat_test_struct_out.lh5"
-    cli.lh5concat(["--output", outfile, "--", infile1, infile2])
+    cli.lh5concat_cli(["--output", outfile, "--", infile1, infile2])
 
     out_stp = store.read("stp", outfile)[0]
     assert out_stp.attrs["datatype"] == "struct{x}"
